Improve shared lock implementation
During lock creation:
Shared locks now have a reservation directory named similar to the exclusive
directories, instead of just 'shared'.  This fixes a race condition where
a process giving up a lock could remove the 'shared' directory successfully.
Meanwhile another wants the shared lock, recreates 'shared', drops it's file
in there, sees the symlink already points to 'shared' and assummes it has
the lock.  Then the first process continues by removing the symlink.  Now
process 2 thinks it has the lock, but it really doesn't.

The fix is if there is any 'shared-*' directory we should use it, otherwise
make a new one.  Now, process 2 will see that the symlink exists, but does
not point to the reservation directory it created, and knows it does not have
the lock.

During unlock:
For exclusive locks, we do the safer method of removing the lock symlink
first, then remove the reservation directory.  That way if we crash before
removing the directory, the lock is still successfully freed.

Shared locks must use the unsafer method of attempting to remove the
reservation directory first, then the symlink of rmdir was successful.

Refuse to free the lock if the current PID does not match the PID used
during create

+++ FileLock.pm
@@ -40,11 +40,15 @@ sub create {
 
     my $self = { style => $params{'style'},
                  resource_lock_dir => $resource_lock_dir,
+                 pid => $$,
                };
     bless $self, $class;
 
     if ($params{'style'} eq 'sh') {
-        $self->{'reservation_dir'} = $resource_lock_dir . 'shared/';
+        #$self->{'reservation_dir'} = $resource_lock_dir . 'shared/';
+        unless ( ($self->{'reservation_dir'}) = (glob($resource_lock_dir . "shared-*/"))[-1] ) {
+            mkdir $self->{'reservation_dir'} = $resource_lock_dir . sprintf('shared-%s-pid%d-%d/',$ENV{'HOST'},$$,time());
+        } 
         $self->{'reservation_file'} = $self->{'reservation_dir'} .
                                   sprintf('%s-pid%d-%d',
                                           $ENV{'HOST'},
@@ -53,7 +57,7 @@ sub create {
     } else {
         # exclusive
         $self->{'reservation_dir'} = $resource_lock_dir .
-                                 sprintf('writer-%s-pid%d-%d/',
+                                 sprintf('excl-%s-pid%d-%d/',
                                          $ENV{'HOST'},
                                          $$,
                                          time());
@@ -107,22 +111,36 @@ sub create {
 sub unlock {
     my $self = shift;
 
-    # shared locks, first remove their file inside the shared/ directory
+    # After a fork(), only the parent should be allowed to unlock?
+    return unless $self->{'pid'} == $$; 
+
     if ($self->{'style'} eq 'sh') {
+        # shared locks, first remove their file inside the shared/ directory
         unlink $self->{'reservation_file'};
-    }
 
-    # Next, remove the lock directory.  
-    # For ex locks, this will always succeed
-    # for sh locks, this will fail if another also has a sh lock
-    my $rv = rmdir $self->{'reservation_dir'};
-    if ($rv) {
-        # If directory was removed, then the lock can be declared free
-        # remove the symlink
+        # Remove the reservation directory
+        my $rv = rmdir $self->{'reservation_dir'};
+
+        # There's a tiny window here where the lock symlink exists, but points to
+        # a non-existent shared reservation directory.  Hope we don't crash and
+        # leave the lock hanging.  There's probably not much we can do to prevent this.
+        
+        # If that worked, we were the last shared lock, remove the lock symlink
+        if ($rv) {
+            unlink $self->{'symlink'} if ($self->{'symlink'});
+        }
+
+    } else {
+        # excl locks
+        
+        # We can safely remove the lock symlink first, giving up the lock
         unlink $self->{'symlink'} if ($self->{'symlink'});
+
+        # Remove our dir to clean up
+        rmdir $self->{'reservation_dir'};
     }
 
-    rmdir $self->{'resource_lock_dir'};
+    rmdir $self->{'resource_lock_dir'};  # If we're the last for this resource, clean up
 
     # make ourselves invalid
     delete $self->{$_} foreach keys %$self;
